+++ release/line-chart/line.component.d.ts
@@ -1,10 +1,14 @@
-import { EventEmitter, ElementRef } from '@angular/core';
-export declare class LineComponent {
+import { EventEmitter, OnChanges, ElementRef, SimpleChanges } from '@angular/core';
+export declare class LineComponent implements OnChanges {
     private element;
     path: any;
     stroke: any;
     data: any;
     fill: string;
     select: EventEmitter<{}>;
+    initialized: boolean;
+    initialPath: string;
     constructor(element: ElementRef);
+    ngOnChanges(changes: SimpleChanges): void;
+    animateToCurrentForm(): void;
 }